Add --ignore-wan-has-ip option
There are now to options for checking whether the interface is up.

The newly added option just checks that the interface has an IP address
assigned. This is a very quick/lightweight check.

The previous option has been renamed to `--ignore-wan-ping-gateway`.
This option requires fork/exec'ing two external processes, first to
determine the IP address of the default gateway, then to ping it to see
whether it is up.

Before even trying `--ignore-wan-ping-gateway`, the first check is
performed, so the prior option implies the new option.

Regardless, the result is still cached for 75 seconds.