Fix icmp_chksum
Truncating the sum to 32 bits before computing the one's complement sum
would compute the wrong sum for large packets. Also renamed it to
ipchecksum since it's suitable for IP, UDP, and TCP and simplified the
code a bit. It also turns out, per RFC 1071, that we don't have to
byte-swap the entire packet. Instead, we can just byte-swap the
checksum at the end of the computation.