Resolve issue where --ping-gateway may choose wrong address
The code implementing the --ping-gateway option previously assumed
there was only a single default route. This is not always the case.
Now the code looks for a gateway (next hop) address restricted to the
if_wan interface. Further, if there is more than one such route, it
uses the next hop of the route with the widest netmask.

Fixes https://github.com/jaysoffian/eap_proxy/issues/10