Harden eap_proxy.sh script for use on USG
With a USG 3P a reboot of the device requires a subsequent reboot of the AT&T Router Gateway (RG) or unplugging/plugging the Ethernet cable on the USG to the AT&T Router Gateway (RG), which both require physical access. This is not ideal as the bypass setup with a USG would not survive a regular reboot. 
Manually setting the interface towards the AT&T ONT with the MAC address of the AT&T Router Gateway (RG) will solve this problem.